requirements.txt
streamlit==1.29.0
openai==1.3.7
llama-index==0.9.12
llama-cpp-python
python-dotenv==1.0.0
setuptools